devide to ImageModel
