+++ app/src/main/java/com/example/finalproject/model/UserModel.java
@@ -1,0 +1,4 @@
+package com.example.finalproject.model;
+
+public class UserModel {
+}
